Added somes classes to define a path that can be drawn or followed by a sprite and text on path
TOGLCPath: a class that encapsulate an array of TPointF to manage a path
TOGLCPathDrawing: a class to draw a path on the scene with color, line width. It support users dotted line.
TSpriteOnPath: define a sprite that can moves along a path
TFreeTextOnPath: to write a text along a path